pyproject.toml
[project]
name = "docanalysis"
version = "0.1.0"
description = "Document analysis and relationship tracking using MCP and ChromaDB"
authors = [
    {name = "mendo" },
]
dependencies = [
    "mcp-agent>=0.0.8",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "chromadb>=0.6.3",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.11",
    "click>=8.1.8",
    "anthropic>=0.49.0",
    "pypdf>=5.3.1",
    "pydantic-settings>=2.8.1",
]
requires-python = ">=3.11"

[project.scripts]
docanalysis = "docanalysis.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.hatch.build.targets.wheel]
packages = ["src/docanalysis"]

[tool.hatch.envs.test]
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "reportlab>=4.1.0",
    "python-dotenv>=1.0.1",
] 
